fix: clippy 零警告（PI 常量/冗余闭包/未用引用）

+++ src-tauri/src/audio/wav.rs
@@ -51,7 +51,7 @@ pub fn read_f32_wav(path: &Path) -> Result<(u32, Vec<f32>), AudioError> {
                 .collect::<Result<Vec<_>, _>>()
                 .map_err(|e| AudioError(e.to_string()))?
                 .into_iter()
-                .map(|s| s as f64 / max as f64)
+                .map(|s| s as f64 / max)
                 .map(|s| s as f32)
                 .collect()
         }
